+++ views/helpers/GeolocationMapBrowse.php
@@ -6,9 +6,9 @@ class Geolocation_View_Helper_GeolocationMapBrowse extends Zend_View_Helper_Abst
     {
         if (!$center) {
             $center = array(
-                'latitude'  => (double) get_option('geolocation_default_latitude'),
-                'longitude' => (double) get_option('geolocation_default_longitude'),
-                'zoomLevel' => (double) get_option('geolocation_default_zoom_level')
+                'latitude'  => (float) get_option('geolocation_default_latitude'),
+                'longitude' => (float) get_option('geolocation_default_longitude'),
+                'zoomLevel' => (float) get_option('geolocation_default_zoom_level')
             );
         }
